jobapptracker/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.3
Django==3.2.13
django-allauth==0.49.0
django-crontab==0.7.1
django-workers==0.2.0
djangorestframework==3.13.1
google-api-core==2.5.0
google-api-python-client==2.39.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
google-cloud-pubsub==2.11.0
googleapis-common-protos==1.55.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jsonfield==3.1.0
lxml==4.8.0
msoffcrypto-tool==5.0.0
oauthlib==3.2.0
olefile==0.46
proto-plus==1.20.3
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.1.0
